backend/requirements.txt
# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Pydantic (updated for v2 compatibility)
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic[email]==2.5.0

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Environment
python-dotenv==1.0.0

# AWS
boto3==1.34.0

# Async tasks
celery==5.3.4
redis==5.0.1

# HTTP client
httpx==0.25.2

# Image processing
Pillow==10.1.0

# Video processing
ffmpeg-python==0.2.0
openai-whisper==20231117
moviepy==1.0.3
numpy==2.4.3
openai==1.12.0

# File handling
aiofiles==25.1.0